changed a, b to short,seem ok for mid data

+++ hdf5t/dynamic_2d_array.h
@@ -7,7 +7,7 @@
 //typedef double integer;
 typedef int integer;
 //typedef short sint;
-typedef int sint;
+typedef short sint;
 typedef char cint;
 typedef float real;
 //typedef double real;
